Patents

Patents.ClassificationCPC

Patents.SymbolCPC

Patents.children

Patents.children_recursive

Patents.parents

Patents.parents_recursive

Patents.prune

Patents.symbols

Patents.ClassificationCPC — Type

Struct to represent the CPC patent classification system.

Patents.SymbolCPC — Type

Struct representing a CPC classification symbol which holds an internal id, the level and corresponding label, the code, and the title of the symbol.

Patents.children — Method

children(classification, symbols)

Get the children for each of the symbols in symbols. symbols can be passed as [a vector of] SymbolCPC or Strings.

Patents.children_recursive — Method

children_recursive(classification, symbols)

Traverse the classification tree given by classification starting at symbols until a sink node is reached and return the visited symbols as a SubClassificationCPC.

Patents.parents — Method

parents(classification, symbols)

Get the parents for each of the symbols in symbols. symbols can be passed as [a vector of] SymbolCPC or Strings.

Patents.parents_recursive — Method

parents_recursive(classification, symbols)

Traverse the classification tree given by classification starting at symbols until a source node is reached and return the visited symbols as a SubClassificationCPC.

Patents.prune — Method

prune(classification, toplevel=1, bottomlevel=5)

Prune the classification tree given by classification to only include nodes in and between toplevel and bottomlevel.

Patents.symbols — Method

symbols(classification)

Obtain an array with SymbolsCPC contained in patent classification object classification. A struct subtyping LevelCPC (e.g. SubgroupCPC) can be passed as a second argument to only obtain symbols of that kind.