Patents

Patents.citationgraph

Patents.normalized_citations

Patents.citationgraph — Method

citationgraph(families)

Return a SimpleDiGraph representing the network of citations among families. An edge from node i to node j indicates that j cited i and will be included in the output graph iff the earliest filing of j is after that of i.

Patents.normalized_citations — Method

normalized_citations(families)

Compute the number of citations for all families in families, normalized relative to the mean citation count of all families in the collection with the same earliest filing year.